+++ package.json
@@ -30,6 +30,7 @@
     "classnames": "^2.5.1",
     "clsx": "^2.1.1",
     "dayjs": "^1.11.13",
+    "formik": "^2.4.6",
     "framer-motion": "^11.3.12",
     "lucide-react": "^0.414.0",
     "measure": "^0.1.1",
@@ -44,7 +45,8 @@
     "styled-components": "^6.1.12",
     "swiper": "^11.1.9",
     "tailwind-merge": "^2.4.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "yup": "^1.4.0"
   },
   "devDependencies": {
     "@types/node": "^20",
